+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/FunctionValidator.java
@@ -41,6 +41,7 @@ import io.ballerina.compiler.syntax.tree.SyntaxKind;
 import io.ballerina.projects.plugins.SyntaxNodeAnalysisContext;
 import io.ballerina.tools.diagnostics.Location;
 
+import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -59,6 +60,8 @@ import static io.ballerina.stdlib.graphql.compiler.Utils.updateContext;
  */
 public class FunctionValidator {
     private final Set<ClassSymbol> visitedClassSymbols = new HashSet<>();
+    private final List<TypeSymbol> existInputObjectTypes = new ArrayList<>();
+    private final List<TypeSymbol> existReturnTypes = new ArrayList<>();
 
     public void validate(SyntaxNodeAnalysisContext context) {
         ServiceDeclarationNode serviceDeclarationNode = (ServiceDeclarationNode) context.node();
@@ -78,6 +81,8 @@ public class FunctionValidator {
                 }
             }
         }
+        existInputObjectTypes.clear();
+        existReturnTypes.clear();
         if (!resourceFunctionFound) {
             updateContext(context, CompilationError.MISSING_RESOURCE_FUNCTIONS, serviceDeclarationNode.location());
         }
@@ -135,7 +140,9 @@ public class FunctionValidator {
 
     private void validateReturnType(TypeSymbol returnTypeDesc,
                                     Location location, SyntaxNodeAnalysisContext context) {
-        if (returnTypeDesc.typeKind() == TypeDescKind.ANY || returnTypeDesc.typeKind() == TypeDescKind.ANYDATA) {
+        if (existInputObjectTypes.contains(returnTypeDesc)) {
+            updateContext(context, CompilationError.INVALID_RETURN_TYPE_INPUT_OBJECT, location);
+        } else if (returnTypeDesc.typeKind() == TypeDescKind.ANY || returnTypeDesc.typeKind() == TypeDescKind.ANYDATA) {
             updateContext(context, CompilationError.INVALID_RETURN_TYPE_ANY, location);
         } else if (returnTypeDesc.typeKind() == TypeDescKind.UNION) {
             validateReturnTypeUnion(((UnionTypeSymbol) returnTypeDesc).memberTypeDescriptors(), location, context);
@@ -162,6 +169,11 @@ public class FunctionValidator {
             TypeDefinitionSymbol typeDefinitionSymbol =
                     (TypeDefinitionSymbol) typeReferenceTypeSymbol.definition();
             if (typeReferenceTypeSymbol.typeDescriptor().typeKind() == TypeDescKind.RECORD) {
+                if (existInputObjectTypes.contains(typeReferenceTypeSymbol.typeDescriptor())) {
+                    updateContext(context, CompilationError.INVALID_RETURN_TYPE_INPUT_OBJECT, location);
+                } else {
+                    existReturnTypes.add(typeReferenceTypeSymbol.typeDescriptor());
+                }
                 validateRecordFields(context, (RecordTypeSymbol) typeDefinitionSymbol.typeDescriptor(), location);
             }
             validateReturnType(typeDefinitionSymbol.typeDescriptor(), location, context);
@@ -186,7 +198,18 @@ public class FunctionValidator {
                 for (ParameterSymbol param : parameterSymbols) {
                     if (hasInvalidInputParamType(param.typeDescriptor())) {
                         Location inputLocation = getLocation(param, location);
-                        updateContext(context, CompilationError.INVALID_RESOURCE_INPUT_PARAM, inputLocation);
+                        if (param.typeDescriptor().typeKind() == TypeDescKind.TYPE_REFERENCE) {
+                            TypeDefinitionSymbol typeDefinitionSymbol = (TypeDefinitionSymbol)
+                                    ((TypeReferenceTypeSymbol) param.typeDescriptor()).definition();
+                            if (typeDefinitionSymbol.typeDescriptor().typeKind() == TypeDescKind.RECORD) {
+                                updateContext(context, CompilationError.INVALID_RESOURCE_INPUT_PARAM_INPUT_OBJECT,
+                                        inputLocation);
+                            } else {
+                                updateContext(context, CompilationError.INVALID_RESOURCE_INPUT_PARAM, inputLocation);
+                            }
+                        } else {
+                            updateContext(context, CompilationError.INVALID_RESOURCE_INPUT_PARAM, inputLocation);
+                        }
                     }
                 }
             }
@@ -212,6 +235,32 @@ public class FunctionValidator {
         if (inputTypeSymbol.typeKind() == TypeDescKind.TYPE_REFERENCE) {
             if ((((TypeReferenceTypeSymbol) inputTypeSymbol).definition()).kind() == SymbolKind.ENUM) {
                 return false;
+            } else if ((((TypeReferenceTypeSymbol) inputTypeSymbol).definition()).kind() ==
+                    SymbolKind.TYPE_DEFINITION) {
+                TypeDefinitionSymbol typeDefinitionSymbol =
+                        (TypeDefinitionSymbol) ((TypeReferenceTypeSymbol) inputTypeSymbol).definition();
+                if (typeDefinitionSymbol.typeDescriptor().typeKind() == TypeDescKind.UNION) {
+                    return true;
+                } else if (typeDefinitionSymbol.typeDescriptor().typeKind() == TypeDescKind.RECORD) {
+                    if (existReturnTypes.contains(typeDefinitionSymbol.typeDescriptor())) {
+                        return true;
+                    } else {
+                        existInputObjectTypes.add(typeDefinitionSymbol.typeDescriptor());
+                    }
+                    Map<String, RecordFieldSymbol> memberMap = ((RecordTypeSymbol) typeDefinitionSymbol
+                            .typeDescriptor()).fieldDescriptors();
+                    boolean hasInvalidMember = true;
+                    for (RecordFieldSymbol fields : memberMap.values()) {
+                        if (fields.typeDescriptor().typeKind() == TypeDescKind.TYPE_REFERENCE) {
+                            hasInvalidMember =
+                                    hasInvalidInputObjectField((TypeReferenceTypeSymbol) fields.typeDescriptor());
+                        } else {
+                            hasInvalidMember = hasInvalidReturnType(fields.typeDescriptor());
+                        }
+                    }
+                    return hasInvalidMember;
+                }
+                return false;
             }
         }
         return !hasPrimitiveType(inputTypeSymbol);
@@ -220,6 +269,17 @@ public class FunctionValidator {
     private boolean hasInvalidReturnType(TypeSymbol returnTypeSymbol) {
         return returnTypeSymbol.typeKind() == TypeDescKind.MAP || returnTypeSymbol.typeKind() == TypeDescKind.JSON ||
                 returnTypeSymbol.typeKind() == TypeDescKind.BYTE || returnTypeSymbol.typeKind() == TypeDescKind.OBJECT;
+    }
+
+    private boolean hasInvalidInputObjectField(TypeReferenceTypeSymbol typeSymbol) {
+        if (typeSymbol.typeDescriptor().typeKind() == TypeDescKind.RECORD) {
+            if (existReturnTypes.contains(typeSymbol.typeDescriptor())) {
+                return true;
+            }
+            existInputObjectTypes.add(typeSymbol.typeDescriptor());
+            return false;
+        }
+        return hasInvalidReturnType(typeSymbol.typeDescriptor());
     }
 
     private void validateServiceClassDefinition(ClassSymbol classSymbol, Location location,
@@ -283,7 +343,12 @@ public class FunctionValidator {
                         if (!isRecordType(returnType)) {
                             updateContext(context, CompilationError.INVALID_RETURN_TYPE, location);
                         } else {
-                            recordTypes++;
+                            if (existInputObjectTypes.contains(returnType)) {
+                                updateContext(context, CompilationError.INVALID_RETURN_TYPE_INPUT_OBJECT, location);
+                            } else {
+                                existReturnTypes.add(returnType);
+                                recordTypes++;
+                            }
                         }
                     } else if (((TypeReferenceTypeSymbol) returnType).definition().kind() == SymbolKind.CLASS) {
                         ClassSymbol classSymbol = (ClassSymbol) ((TypeReferenceTypeSymbol) returnType).definition();